Updated tests. 1 of 8 are passing so far.
