+++ test/src/Item.svelte
@@ -1,0 +1,5 @@
+<script>
+    let index = 0;
+</script>
+
+<span>{index}</span>
