+++ test/src/utils.js
@@ -1,0 +1,60 @@
+export function indent(node, spaces) {
+	if (node.childNodes.length === 0) return;
+
+	if (node.childNodes.length > 1 || node.childNodes[0].nodeType !== 3) {
+		const first = node.childNodes[0];
+		const last = node.childNodes[node.childNodes.length - 1];
+
+		const head = `\n${spaces}  `;
+		const tail = `\n${spaces}`;
+
+		if (first.nodeType === 3) {
+			first.data = `${head}${first.data}`;
+		} else {
+			node.insertBefore(document.createTextNode(head), first);
+		}
+
+		if (last.nodeType === 3) {
+			last.data = `${last.data}${tail}`;
+		} else {
+			node.appendChild(document.createTextNode(tail));
+		}
+
+		let lastType = null;
+		for (let i = 0; i < node.childNodes.length; i += 1) {
+			const child = node.childNodes[i];
+			if (child.nodeType === 1) {
+				indent(node.childNodes[i], `${spaces}  `);
+
+				if (lastType === 1) {
+					node.insertBefore(document.createTextNode(head), child);
+					i += 1;
+				}
+			}
+
+			lastType = child.nodeType;
+		}
+	}
+}
+
+export function normalize(html) {
+	const div = document.createElement('div');
+	div.innerHTML = html
+		.replace(/<!--.+?-->/g, '')
+        .replace(/<object.+\/object>/g, '')
+        .replace(/\s*svelte-\w+\s*/g, '')
+		.replace(/class="svelte-\w+"\s*/g, '')
+		.replace(/>\s+/g, '>')
+		.replace(/\s+</g, '<');
+
+	indent(div, '');
+
+	div.normalize();
+	return div.innerHTML;
+}
+
+export function sleep(ms) {
+	return new Promise(fulfil => {
+		setTimeout(fulfil, ms);
+	});
+}
